refactor: TrnasactionTemplate  일반 클래스로 변경
