+++ jdbc/src/main/java/com/interface21/transaction/support/TransactionTemplate.java
@@ -1,10 +1,8 @@
-package com.techcourse.service;
+package com.interface21.transaction.support;
 
 import com.interface21.dao.DataAccessException;
 import com.interface21.jdbc.core.ConnectionConsumerWrapper;
-import com.interface21.jdbc.core.ServiceLogic;
 import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.techcourse.config.DataSourceConfig;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Objects;
@@ -12,13 +10,14 @@ import javax.sql.DataSource;
 
 public class TransactionTemplate {
 
-    private static final DataSource DATASOURCE = DataSourceConfig.getInstance();
+    private final DataSource dataSource;
 
-    private TransactionTemplate() {
+    public TransactionTemplate(DataSource dataSource) {
+        this.dataSource = dataSource;
     }
 
-    public static void executeWithTransaction(ServiceLogic logic) {
-        Connection connection = DataSourceUtils.getConnection(DATASOURCE);
+    public void executeWithTransaction(ServiceLogic logic) {
+        Connection connection = DataSourceUtils.getConnection(dataSource);
         try {
             validateConnection(connection);
             connection.setAutoCommit(false);
@@ -28,7 +27,7 @@ public class TransactionTemplate {
             ConnectionConsumerWrapper.accept(connection, Connection::rollback);
             throw new DataAccessException(e);
         } finally {
-            DataSourceUtils.releaseConnection(connection, DATASOURCE);
+            DataSourceUtils.releaseConnection(connection, dataSource);
         }
     }
